// frontend/layout/lib/types.ts
export enum NavigateDirection {
    Up = 0,
    Right = 1,
    Down = 2,
    Left = 3,
}

export enum DropDirection {
    Top = 0,
    Right = 1,
    Bottom = 2,
    Left = 3,
    OuterTop = 4,
    OuterRight = 5,
    OuterBottom = 6,
    OuterLeft = 7,
    Center = 8,
}

export enum FlexDirection {
    Row = "row",
    Column = "column",
}